fix: 인터페이스 통합
