--- apps/ceos/src/components/recruit/interface.ts
@@ -1,11 +1,0 @@
-export interface PassDataInterface {
-  uuid: string;
-  generation: number;
-  email: string;
-  pass: string;
-  name: string;
-  attendanceStatus: boolean;
-  date: string;
-  otDate: string;
-  duration: string;
-}
