+++ apps/ceos/src/components/recruit/interface.tsx
@@ -43,3 +43,15 @@ export type PartName =
   | 'designQuestions'
   | 'frontendQuestions'
   | 'backendQuestions';
+
+export interface PassDataInterface {
+  uuid: string;
+  generation: number;
+  email: string;
+  pass: string;
+  name: string;
+  attendanceStatus: boolean;
+  date: string;
+  otDate: string;
+  duration: string;
+}
